+++ Src/DiffContext.h
@@ -212,6 +212,7 @@ public:
 	std::shared_ptr<SubstitutionList> m_pSubstitutionList; /// list for Substitution Filters
 	std::unique_ptr<PropertySystem> m_pPropertySystem; /**< pointer to Property System */
 	std::vector<std::map<std::vector<uint8_t>, DuplicateInfo>> m_duplicateValues; /**< Number of duplicate hash values */
+	std::vector<String> m_vCurrentlyHiddenItems; /**< The list of currently hidden items */
 
 private:
 	/**