+++ Src/DirView.h
@@ -201,7 +201,6 @@ protected:
 	int m_nEscCloses; /**< Cached value for option for ESC closing window */
 	bool m_bExpandSubdirs;
 	CFont m_font; /**< User-selected font */
-	UINT m_nHiddenItems; /**< Count of items we have hidden */
 	bool m_bTreeMode; /**< `true` if tree mode is on*/
 	DirViewFilterSettings m_dirfilter;
 	clock_t m_compareStart; /**< Starting process time of the compare */
@@ -391,6 +390,8 @@ protected:
 	DECLARE_MESSAGE_MAP()
 	bool OnHeaderBeginDrag(LPNMHEADER hdr, LRESULT* pResult);
 	bool OnHeaderEndDrag(LPNMHEADER hdr, LRESULT* pResult);
+	void HideItems(const std::vector<String>& ItemsToHide);
+	bool IsItemToHide(const String& currentItem, const std::vector<String>& ItemsToHide) const;
 
 private:
 	void Open(CDirDoc *pDoc, const PathContext& paths, DWORD dwFlags[3], FileTextEncoding encoding[3], PackingInfo * infoUnpacker = nullptr);